# apps/docs/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mayarin API",
    "version": "2026-08-11",
    "description": "Create fiat-denominated commerce requests, accept supported crypto assets, and settle merchants in stablecoins. Versioned operations live under `/v1`; buyer-facing checkout, invoice, and health routes are intentionally unversioned."
  },
  "servers": [
    {
      "url": "https://api.mayarin.xyz",
      "description": "Mainnet"
    },
    {
      "url": "https://api-testnet.mayarin.xyz",
      "description": "Testnet"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "secretKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server-side `sk_` key. Grants a subset of the merchant's permissions."
      },
      "publishableOrSecretKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Browser-safe `pk_` key, or a stronger server-side `sk_` key. Identifies the merchant for catalog read and cart checkout; carries no permissions."
      }
    },
    "schemas": {
      "Merchant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "categoryCode": {
            "type": "string"
          }
        },
        "required": ["id", "name", "city", "countryCode"]
      },
...
      "PaymentIntent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Payment intent lifecycle state."
          },
          "merchant": {
            "$ref": "#/components/schemas/Merchant"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "settlementAsset": {
            "type": "string"
          },
          "provider": {
            "type": ["string", "null"]
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentRail"
          },
          "executionPath": {
            "type": ["string", "null"]
          },
          "source": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "merchantReference": {
            "type": ["string", "null"]
          },
          "clearingTransactionId": {
            "type": ["string", "null"]
          },
          "failureReason": {
            "type": ["string", "null"]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "confirmedAt": {
            "type": ["string", "null"]
          },
          "completedAt": {
            "type": ["string", "null"]
          }
        },
        "required": [
          "id",
          "status",
          "merchant",
          "amount",
          "settlementAsset",
          "provider",
          "payment",
          "executionPath",
          "source",
          "metadata",
          "merchantReference",
          "clearingTransactionId",
          "failureReason",
          "createdAt",
          "updatedAt",
          "expiresAt",
          "confirmedAt",
          "completedAt"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              }
            },
            "required": ["code", "message", "retryable"]
          }
        },
        "required": ["error"]
      },
      "Clearing": {
        "type": ["object", "null"],
        "properties": {
          "id": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "description": "Clearing transaction state."
          },
          "provider": {
            "type": "string"
          },
          "providerReference": {
            "type": ["string", "null"]
          },
          "transactionHash": {
            "type": ["string", "null"]
          },
          "sourceAmount": {
            "$ref": "#/components/schemas/Money"
          },
          "settlementAmount": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": ["object", "null"]
              }
            ]
          },
          "fee": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": ["object", "null"]
              }
            ]
          },
          "netAmount": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": ["object", "null"]
              }
            ]
          },
          "rate": {
            "type": ["object", "null"],
            "properties": {
              "from": {
                "type": "string"
              },
              "to": {
                "type": "string"
              },
              "scaledRate": {
                "type": "string"
              },
              "source": {
                "type": "string"
              },
              "lockedAt": {
                "type": "string"
              }
            },
            "required": ["from", "to", "scaledRate", "source", "lockedAt"]
          },
          "failure": {
            "type": ["object", "null"],
            "properties": {
              "reason": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "at": {
                "type": "string"
              }
            },
            "required": ["reason", "code", "at"]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "provider",
          "providerReference",
          "transactionHash",
          "sourceAmount",
          "settlementAmount",
          "fee",
          "netAmount",
          "rate",
          "failure",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Payment": {
        "type": "object",
        "properties": {
          "paymentIntent": {
            "$ref": "#/components/schemas/PaymentIntent"
          },
          "clearing": {
            "$ref": "#/components/schemas/Clearing"
          },
          "deposit": {
            "$ref": "#/components/schemas/Deposit"
          },
          "timeline": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "sequence": {
                  "type": "integer"
                },
                "state": {
                  "type": "string"
                },
                "from": {
                  "type": ["string", "null"]
                },
                "occurredAt": {
                  "type": "string"
                }
              },
              "required": ["sequence", "state", "from", "occurredAt"]
            }
          }
        },
        "required": ["paymentIntent", "clearing", "deposit", "timeline"]
      },
...
      "PaymentLink": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "description": "fixed, open, or catalog"
          },
          "merchant": {
            "$ref": "#/components/schemas/Merchant"
          },
          "amount": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": ["object", "null"]
              }
            ]
          },
          "currency": {
            "type": ["string", "null"]
          },
          "lines": {
            "type": ["array", "null"],
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                }
              },
              "required": ["productId", "quantity"]
            }
          },
          "title": {
            "type": ["string", "null"]
          },
          "merchantReference": {
            "type": ["string", "null"]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "url": {
            "type": "string"
          },
          "payable": {
            "type": "boolean"
          },
          "listed": {
            "type": "boolean",
            "description": "Whether it appears in the public x402 payable index."
          },
          "expiresAt": {
            "type": ["string", "null"]
          },
          "disabledAt": {
            "type": ["string", "null"]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "merchant",
          "amount",
          "currency",
          "lines",
          "title",
          "merchantReference",
          "metadata",
          "url",
          "payable",
          "listed",
          "expiresAt",
          "disabledAt",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/payments/{id}": {
      "get": {
        "tags": ["Payments"],
        "operationId": "getPayment",
        "summary": "Get payment status",
        "description": "Accepts a payment intent id (`pi_…`) or clearing transaction id (`clr_…`).",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment, clearing state, deposit, and timeline.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          },
          "404": {
            "description": "No payment with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/payment-links": {
      "get": {
        "tags": ["Payment links"],
        "operationId": "listPaymentLinks",
        "summary": "List payment links",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "merchantId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment links.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentLink"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Payment links"],
        "operationId": "createPaymentLink",
        "summary": "Create a payment link",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Stable unique key for this logical write. Keep it unchanged when retrying."
            },
            "required": true,
            "description": "Stable unique key for this logical write. Keep it unchanged when retrying.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": ["fixed", "open", "catalog"]
                  },
                  "merchant": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "minLength": 1
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1
                      },
                      "city": {
                        "type": "string",
                        "minLength": 1
                      },
                      "countryCode": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 2
                      },
                      "categoryCode": {
                        "type": "string"
                      }
                    },
                    "required": ["id", "name", "city", "countryCode"]
                  },
                  "amount": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "string"
                      },
                      "asset": {
                        "type": "string",
                        "enum": [
                          "IDR",
                          "USD",
                          "SGD",
                          "THB",
                          "MYR",
                          "USDC",
                          "USDT",
                          "EURC",
                          "ETH",
                          "BTC"
                        ]
                      }
                    },
                    "required": ["amount", "asset"]
                  },
                  "currency": {
                    "type": "string",
                    "enum": [
                      "IDR",
                      "USD",
                      "SGD",
                      "THB",
                      "MYR",
                      "USDC",
                      "USDT",
                      "EURC",
                      "ETH",
                      "BTC"
                    ]
                  },
                  "lines": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "quantity": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        }
                      },
                      "required": ["productId", "quantity"]
                    },
                    "minItems": 1
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "merchantReference": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "listed": {
                    "type": "boolean"
                  }
                },
                "required": ["kind", "merchant"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment link created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLink"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payment-links/{id}": {
      "get": {
        "tags": ["Payment links"],
        "operationId": "getPaymentLink",
        "summary": "Get a payment link",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment link.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLink"
                }
              }
            }
          },
          "404": {
            "description": "No link with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payment-links/{id}/disable": {
      "post": {
        "tags": ["Payment links"],
        "operationId": "disablePaymentLink",
        "summary": "Disable a payment link",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Stable unique key for this logical write. Keep it unchanged when retrying."
            },
            "required": true,
            "description": "Stable unique key for this logical write. Keep it unchanged when retrying.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Disabled payment link.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLink"
                }
              }
            }
          },
          "404": {
            "description": "No link with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payment-links/{id}/list": {
      "post": {
        "tags": ["Payment links"],
        "operationId": "listPaymentLink",
        "summary": "List a payment link in the public payable index",
        "description": "Opts the link into the public x402 payable index, where agents discover and pay it. The link stays payable by anyone holding its URL either way; listing only decides whether a discovery reader is shown it.",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Listed payment link.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLink"
                }
              }
            }
          },
          "404": {
            "description": "No link with that id, or not this merchant's.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payment-links/{id}/unlist": {
      "post": {
        "tags": ["Payment links"],
        "operationId": "unlistPaymentLink",
        "summary": "Remove a payment link from the public payable index",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Unlisted payment link.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLink"
                }
              }
            }
          },
          "404": {
            "description": "No link with that id, or not this merchant's.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payment-links/{id}/checkout": {
      "post": {
        "tags": ["Payment links"],
        "operationId": "checkoutPaymentLink",
        "summary": "Check out a payment link",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Stable unique key for this logical write. Keep it unchanged when retrying."
            },
            "required": true,
            "description": "Stable unique key for this logical write. Keep it unchanged when retrying.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "settlementAsset": {
                    "type": "string",
                    "enum": [
                      "IDR",
                      "USD",
                      "SGD",
                      "THB",
                      "MYR",
                      "USDC",
                      "USDT",
                      "EURC",
                      "ETH",
                      "BTC"
                    ]
                  },
                  "provider": {
                    "type": "string",
                    "minLength": 1
                  },
                  "payment": {
                    "type": "object",
                    "properties": {
                      "asset": {
                        "type": "string",
                        "enum": [
                          "IDR",
                          "USD",
                          "SGD",
                          "THB",
                          "MYR",
                          "USDC",
                          "USDT",
                          "EURC",
                          "ETH",
                          "BTC"
                        ]
                      },
                      "chain": {
                        "type": "string",
                        "enum": [
                          "base",
                          "base-sepolia",
                          "arbitrum",
                          "arbitrum-sepolia",
                          "robinhood-testnet",
                          "arc-testnet"
                        ]
                      },
                      "payerAddress": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{40}$"
                      }
                    },
                    "required": ["asset", "chain"]
                  },
                  "executionPath": {
                    "type": "string",
                    "enum": ["deposit-match", "on-chain-contract", "x402"]
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "merchantReference": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "ttlSeconds": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "amount": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "string"
                      },
                      "asset": {
                        "type": "string",
                        "enum": [
                          "IDR",
                          "USD",
                          "SGD",
                          "THB",
                          "MYR",
                          "USDC",
                          "USDT",
                          "EURC",
                          "ETH",
                          "BTC"
                        ]
                      }
                    },
                    "required": ["amount", "asset"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment intent created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntent"
                }
              }
            }
          },
          "404": {
            "description": "No link with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/checkout/pay/{intentId}": {
      "get": {
        "tags": ["Unversioned"],
        "operationId": "getCheckoutPayPage",
        "summary": "Hosted pay page for one intent",
        "description": "Intentionally unversioned. Renders the pay flow for a single payment intent.",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "intentId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Pay HTML page.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}